Modified database Book.series attribute
I realized that putting series as "Lord of the Rings:1" composite data was a bad idea for future use of this field. So I split the attribute into two: series and seriesNum.

Because this added an extra attribute, it may modify the previous code that reference particular column index values for books.

I updated my index values in the HomeController class.