Don't worry the Profile Page still works!!
Currently working on swapping the image view so that the user's profile pic is updated accordingly, but it is not doing such thing and will make the profile page appear as though it no longer has the liked books and wishlist fields. However, if you comment out the Image object declaration and the imageView.setImage then it will go back to normal. Just commiting to save what I have thus far.

Also, caught some bugs that were displaying the text fields incorrectly when going back and forth between the settings page and the profile page after saving the changes.

Then, implemented a randomizer so every new account created gets a random number between 0-7 to give them their initial profile pic from a newly declared array holding the names of the images/pngs corresponding to the array index.

Finally, got the settings icon added to the repo and placed in front of the settings clickable link (the cog isn't clickable, but can be if y'all want).