One last BIG push!!!
Finally, got the profile pics to update using an arrow system. Both methods, prevWorm(goes left) and nextWorm(go right) take into account the bounds of the array that stores the file names of the pngs.

Adjusted some FXML to allow for more seamless traversal.

Added the left and right arrows to the repo which like stated above were used to update the worm icons.